ACDevice.py

import json
import paho.mqtt.client as mqtt

from Topics import Topic


HOST = "localhost"
PORT = 1883
    
class AC_Device():
    
    _MIN_TEMP = 18  
    _MAX_TEMP = 32  

    def __init__(self, device_id, room):
        self._device_id = device_id
        self._room_type = room
        self._temperature = 22
        self._device_type = "AC"
        self._device_registration_flag = False
        self._topics = set()
        self.client = mqtt.Client(self._device_id)  
        self.client.on_connect = self._on_connect  
        self.client.on_message = self._on_message  
        self.client.on_disconnect = self._on_disconnect  
        self.client.connect(HOST, PORT, keepalive=60)  
        self.client.loop_start()  
        self._register_device(self._device_id, self._room_type, self._device_type)
        self._switch_status = "OFF"

    # calling registration method to register the device
    def _register_device(self, device_id, room_type, device_type):
        # During creation of the device, implement a call to the server to register itself
        payload = {
            'device_id': device_id,
            'device_type': device_type,
            'room_type': room_type
        }
        self.client.publish(
            Topic.REGISTRATION_REQ,
            payload=json.dumps(payload)
        )

    # Connect method to subscribe to various topics. 
    def _on_connect(self, client, userdata, flags, result_code):
        # Subscribe to topics once the device is connected to MQTT broker
        self._subscribe_topics()

    # method to process the recieved messages and publish them on relevant topics 
    # this method can also be used to take the action based on received commands
    def _on_message(self, client, userdata, msg): 
        if msg.topic == f"{Topic.REGISTRATION_ACK}/{self._device_id}":
            print(f"{self._device_id} is connected")

    def _on_disconnect(self, client, userdata, rc):
        pass

    # Getting the current switch status of devices 
    def _get_switch_status(self):
        return self._switch_status

    # Setting the the switch of devices
    def _set_switch_status(self, switch_state):
        pass

    # Getting the temperature for the devices
    def _get_temperature(self):
        pass        

    # Setting up the temperature of the devices
    def _set_temperature(self, temperature):
        pass

    def _subscribe_topics(self):
        self._topics = {
            f"{Topic.REGISTRATION_ACK}/{self._device_id}",
        }
        for topic in self._topics:
            self.client.subscribe(topic)
        print(f"{self._device_id} subscribes to {self._topics}")
    